requirements.txt
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
docx2txt==0.8
decorator==4.4.2
Flask==2.0.1
google-api-core==1.29.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-language==2.0.0
google-cloud-speech==2.4.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
libcst==0.3.19
MarkupSafe==2.0.1
moviepy==1.0.3
mypy-extensions==0.4.3
numpy==1.20.3
opencv-python==4.5.5.64
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
proglog==0.1.9
proto-plus==1.18.1
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.25.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==6.0.1
requests==2.25.1
rsa==4.7.2
# scikit-learn==1.0.2
# scipy==1.6.3
six==1.16.0
threadpoolctl==2.1.0
tqdm==4.61.0
transformers==4.12.3
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.5
Werkzeug==2.0.1
torch~=1.5.1
msrest~=0.6.21
librosa~=0.9.1